(PUP-3930) Optimize failed_dependencies?
I profiled `puppet agent --test` on one of my servers using stackprof
[1] on Ruby 2.1.

Something like 30% or more of the time was spent in
`Puppet::Graph::SimpleGraph#upstream_from_vertex`, called from
`Puppet::Transaction#failed_dependencies?`

It turns out that, while evaluating the resource graph, when considering
a resource, we look at *the complete set of transitive dependencies* to
evaluate whether any of them have failed. This is hugely expensive, and
moreover, is wasted work in the success case where no resources fail.

Instead of all that work, this patch pushes the work to the *failed*
nodes; When a node fails, we transitively walk the dependents of that
node, and mark them as having failed dependencies, and then check that
flag directly when considering whether to skip a node later.

On my test system, this patch drops `puppet agent --test` runtime from
about 40s to about 23s.

[1] https://github.com/tmm1/stackprof